examples/argocd/generate_code/pyproject.toml

# Copyright 2025 CNOE
# SPDX-License-Identifier: Apache-2.0
# Generated by CNOE OpenAPI MCP Codegen tool

[project]
name = "agent_argocd"
version = "0.1.0"
description = "LangGraph agent for argocd MCP tools"
license = "Apache-2.0"
readme = "README.md"
authors = [
    { name = "Sri Aradhyula", email = "sraradhy@cisco.com" },
]
requires-python = ">=3.13,<4.0"
dependencies = [

        "a2a-sdk>=0.2",
        "httpx>=0.28",
        "agntcy-acp>=1.3.2",
        "click>=8.2.0",
        "langchain-anthropic>=0.3.13",
        "langchain-core>=0.3.60",
        "langchain-google-genai>=2.1.4",
        "langchain-mcp-adapters>=0.1.9",
        "langchain-openai>=0.3.17",
        "langfuse>=3.2.0",
        "langchain>=0.3.27",
        "langgraph>=0.4.5",
        "uv",
        "rich (>=14.0.0,<15.0.0)",
        "sseclient (>=0.0.27,<0.0.28)",
        "cnoe-agent-utils>=0.3",
        "fastmcp>=2.11.1",
          
]

[tool.hatch.build.targets.wheel]
# include every top-level package in the current directory
packages = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.poetry.scripts]

agent_argocd_a2a = "protocol_bindings.a2a_server:main"


[tool.ruff]
line-length = 140
indent-width = 2

[tool.ruff.lint]
select = ["E", "F"]
ignore = ["F403"]

[tool.hatch.metadata]
allow-direct-references = true